fix(plugin-content-pdf-fragment): fine tune #viewrect open parameter support
When adding `#viewrect` on the fly, using the conditional wrapper causes
the `Page` component to re-render the PDF page. While the PDF is
rerendering and if SizedLoader doesn't have a good default size, `div.scrollTo`
would fail, since the div's child doesn't have the expected size yet. It
will have the size only after the PDF page is re-rendered.
Solution: don't use the conditional wrapper. Just let it be there and
crop it when `#viewrect` is used.